整理代码2
