+++ DataKitsServer/src/main/java/io/github/xuyao5/datakitsserver/job/MySQL2EsDemoJob.java
@@ -1,6 +1,5 @@
 package io.github.xuyao5.datakitsserver.job;
 
-import com.github.shyiko.mysql.binlog.jmx.BinaryLogClientMXBean;
 import io.github.xuyao5.datakitsserver.configuration.EsKitsConfig;
 import io.github.xuyao5.dkl.eskits.service.MySQL2EsService;
 import io.github.xuyao5.dkl.eskits.service.config.MySQL2EsConfig;
@@ -9,8 +8,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.elasticsearch.client.RestHighLevelClient;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-
-import java.util.concurrent.Future;
 
 @Slf4j
 @Component("mySQL2EsDemoJob")
@@ -25,7 +22,7 @@ public final class MySQL2EsDemoJob implements Runnable {
     @SneakyThrows
     @Override
     public void run() {
-        Future<BinaryLogClientMXBean> execute = new MySQL2EsService(esClient, esKitsConfig.getEsBulkThreads()).execute(MySQL2EsConfig.of("BinlogTest", "root", "123456", new String[]{"MyTable"}));
+        new MySQL2EsService(esClient, "BinlogTest", "root", "123456", esKitsConfig.getEsBulkThreads()).execute(MySQL2EsConfig.of(new String[]{"MyTable"}));
         System.in.read();
     }
 }
