+++ EsKitsLibrary/src/main/java/io/github/xuyao5/dkl/eskits/service/config/MySQL2EsConfig.java
@@ -11,12 +11,5 @@ import lombok.Data;
 @Data(staticConstructor = "of")
 public final class MySQL2EsConfig {
 
-    private final String schema;
-    private final String username;
-    private final String password;
-    private final String[] table;
-
-    private String hostname;
-    private int port;
-    private long timeout;
+    private final String[] tables;
 }
